chore: add utils functions for sharing
